MULTISITE-2539: Implement the correct hook.
As part of MULTISITE-173 a hook was added to Workbench Moderation to allow to
override to which states a user can override a node. A few weeks later this
functionality was added to Workbench Moderation as part of issue #1108502.

This replaces the original alter hook and its implementation with the one that
was actually included in Workbench Moderation.