NEXTEUROPA-3092: Updated patch.
This patch file consisted of the fix for issue #1605252, as well as a hack
that was added in MULTISITE-550. Issue #1605252 has in the meanwhile been fixed
upstream so this patch does not apply any more.

This commit only fixes the conflict with the new upstream release. The hack from
MULTISITE-550 is still there.