NEXTEUROPA-3297: Updated documentation.
The concurrency was originally increased to work around a bug in Drush but this
has since been fixed. But a higher concurrency also greatly improves the build
time. The build is around 2.5x-3x faster with a concurrency of 10. Going any
higher than this doesn't make it any faster though.