NEXTEUROPA-4216: Decrease minimum stability to dev.
The latest stable version of Coder throws warnings which are solved in the
development version. However the development version of Coder depends on the
development version of PHP CodeSniffer. In order to be able to install this we
need to temporarily lower our minimum stability to "dev".

Once new versions of PHP CodeSniffer and Coder are released this commit should
be reverted.