Feature: Update existing users with the same email instead of creating new ones.
Imagine the following use case scenario:
- A user has registered on the drupal website by filling a drupal registration form.
At some point, the user attempts to use the the EU login and enters the same email with the one that is registered on drupal.
Once authenticated, a new account will be created for the user instead of syncing the info with the existing one (assuming that drupal and ecas usernames are different).
That is happening because in the ecas_login_check() function, users are loaded by their username. I suggest that
in case user_load_by_name() fails, we try loading the user by mail and if that succeeds, register the user to the authmap table and update the username.